Add tests for the default action behaviour
