+++ tests/unit/actions/test_steps.py
@@ -33,36 +33,6 @@ ALL_STEPS = {
 }
 
 
-def test_default_invalid_init():
-    with pytest.raises(TypeError):
-        default.init()  # pylint: disable=no-value-for-parameter
-
-
-def test_default_returns_callable_without_data():
-    callable_object = default.init(jira_project_key="")
-    assert callable_object
-    with pytest.raises(TypeError) as exc_info:
-        assert callable_object()
-
-    assert "missing 1 required positional argument: 'context'" in str(exc_info.value)
-
-
-def test_default_returns_callable_with_data(
-    context_create_example: ActionContext, mocked_jira, mocked_bugzilla
-):
-    sentinel = mock.sentinel
-    mocked_jira.create_issue.return_value = {"key": "k"}
-    mocked_jira.create_or_update_issue_remote_links.return_value = sentinel
-    mocked_bugzilla.get_bug.return_value = context_create_example.bug
-    callable_object = default.init(jira_project_key=context_create_example.jira.project)
-
-    handled, details = callable_object(context=context_create_example)
-
-    assert handled
-    assert details["responses"][0] == {"key": "k"}
-    assert details["responses"][1] == sentinel
-
-
 def test_created_public(
     context_create_example: ActionContext, mocked_jira, mocked_bugzilla
 ):
